Revert 'Move mutator implementations out to a standalone service'
