+++ build-system/tasks/presubmit-checks.js
@@ -224,14 +224,6 @@ const forbiddenTerms = {
       'testing/iframe.js',
     ],
   },
-  'installMutatorServiceForDoc': {
-    message: privateServiceFactory,
-    whitelist: [
-      'src/inabox/inabox-services.js',
-      'src/service/core-services.js',
-      'src/service/mutator-impl.js',
-    ],
-  },
   'installPerformanceService': {
     message: privateServiceFactory,
     whitelist: [
@@ -239,14 +231,6 @@ const forbiddenTerms = {
       'src/amp-shadow.js',
       'src/inabox/amp-inabox.js',
       'src/service/performance-impl.js',
-    ],
-  },
-  'installResourcesServiceForDoc': {
-    message: privateServiceFactory,
-    whitelist: [
-      'src/inabox/inabox-services.js',
-      'src/service/core-services.js',
-      'src/service/resources-impl.js',
     ],
   },
   'installStorageServiceForDoc': {
@@ -298,6 +282,15 @@ const forbiddenTerms = {
       'src/service/resources-impl.js',
       'src/service/viewport/viewport-impl.js',
       'src/service/vsync-impl.js',
+    ],
+  },
+  'installResourcesServiceForDoc': {
+    message: privateServiceFactory,
+    whitelist: [
+      'src/inabox/inabox-services.js',
+      'src/service/core-services.js',
+      'src/service/resources-impl.js',
+      'src/service/standard-actions-impl.js',
     ],
   },
   'installXhrService': {
@@ -579,7 +572,7 @@ const forbiddenTerms = {
   },
   '\\.schedulePass\\(': {
     message: 'schedulePass is heavy, think twice before using it',
-    whitelist: ['src/service/mutator-impl.js', 'src/service/resources-impl.js'],
+    whitelist: ['src/service/resources-impl.js'],
   },
   '\\.requireLayout\\(': {
     message:
@@ -856,7 +849,6 @@ const forbiddenTerms = {
       'test/unit/test-mode.js',
       'test/unit/test-motion.js',
       'test/unit/test-mustache.js',
-      'test/unit/test-mutator.js',
       'test/unit/test-object.js',
       'test/unit/test-observable.js',
       'test/unit/test-pass.js',
