+++ extensions/amp-story-auto-ads/0.1/story-ad-button-text-fitter.js
@@ -31,8 +31,8 @@ export class ButtonTextFitter {
    * @param {!../../../src/service/ampdoc-impl.AmpDoc} ampdoc
    */
   constructor(ampdoc) {
-    /** @const @private {!../../../src/service/mutator-interface.MutatorInterface} */
-    this.mutator_ = Services.mutatorForDoc(ampdoc);
+    /** @private @const {!../../../src/service/resources-interface.ResourcesInterface} */
+    this.resources_ = Services.resourcesForDoc(ampdoc);
 
     /** @private {!Document} */
     this.doc_ = ampdoc.win.document;
@@ -40,7 +40,7 @@ export class ButtonTextFitter {
     /** @private {!Element} */
     this.measurer_ = this.doc_.createElement('div');
 
-    this.mutator_.mutateElement(this.measurer_, () => {
+    this.resources_.mutateElement(this.measurer_, () => {
       this.doc_.body.appendChild(this.measurer_);
       setStyles(this.measurer_, {
         position: 'absolute',
@@ -62,7 +62,7 @@ export class ButtonTextFitter {
    */
   fit(pageElement, container, content) {
     let success = false;
-    return this.mutator_
+    return this.resources_
       .mutateElement(container, () => {
         this.measurer_.textContent = content;
         const fontSize = calculateFontSize(
