+++ src/inabox/inabox-mutator.js
@@ -15,7 +15,6 @@
  */
 
 import {Services} from '../services';
-import {registerServiceBuilderForDoc} from '../service';
 
 /**
  * @implements {../service/mutator-interface.MutatorInterface}
@@ -23,10 +22,11 @@ import {registerServiceBuilderForDoc} from '../service';
 export class InaboxMutator {
   /**
    * @param {!../service/ampdoc-impl.AmpDoc} ampdoc
+   * @param {!../service/resources-interface.ResourcesInterface} resources
    */
-  constructor(ampdoc) {
+  constructor(ampdoc, resources) {
     /** @const @private {!../service/resources-interface.ResourcesInterface} */
-    this.resources_ = Services.resourcesForDoc(ampdoc);
+    this.resources_ = resources;
 
     /** @private @const {!../service/vsync-impl.Vsync} */
     this.vsync_ = Services./*OK*/ vsyncFor(ampdoc.win);
@@ -100,11 +100,4 @@ export class InaboxMutator {
       },
     });
   }
-}
-
-/**
- * @param {!../service/ampdoc-impl.AmpDoc} ampdoc
- */
-export function installInaboxMutatorServiceForDoc(ampdoc) {
-  registerServiceBuilderForDoc(ampdoc, 'mutator', InaboxMutator);
 }
